+++ libindi/drivers/telescope/lx200basic.h
@@ -31,7 +31,7 @@ class LX200Basic : public INDI::Telescope
 
  virtual const char *getDefaultName();
  virtual bool Connect();
- virtual bool Connect(const char *port, uint16_t baud);
+ virtual bool Connect(const char *port, uint32_t baud);
  virtual bool Disconnect();
  virtual bool ReadScopeStatus();
  virtual void ISGetProperties(const char *dev);
