+++ libindi/drivers/telescope/lx200generic.h
@@ -36,7 +36,7 @@ class LX200Generic: public INDI::Telescope, public INDI::GuiderInterface
 
     virtual const char *getDefaultName();
     virtual bool Connect();
-    virtual bool Connect(const char *port, uint16_t baud);
+    virtual bool Connect(const char *port, uint32_t baud);
     virtual bool Disconnect();
     virtual bool ReadScopeStatus();
     virtual void ISGetProperties(const char *dev);
